requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.2.0=eigen
abseil-cpp=20211102.0=he9d5cce_0
absl-py=1.3.0=py310hecd8cb5_0
aiohttp=3.8.1=py310hca72f7f_1
aiosignal=1.2.0=pyhd3eb1b0_0
altair=4.2.0=pypi_0
appnope=0.1.2=py310hecd8cb5_1001
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py310hecd8cb5_0
attrs=22.1.0=py310hecd8cb5_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py310hecd8cb5_0
bottleneck=1.3.5=py310h4e76f89_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py310hca72f7f_1002
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.10.11=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.9.24=py310hecd8cb5_0
cffi=1.15.1=py310hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310hecd8cb5_0
commonmark=0.9.1=pypi_0
cryptography=38.0.1=py310hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py310he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py310hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
fftw=3.3.9=h9ed2024_1
flatbuffers=2.0.0=h23ab428_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.2.0=py310hca72f7f_1
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=haf1e3a3_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.46.1=h067a048_0
grpcio=1.42.0=py310ha29bfda_0
h5py=3.7.0=py310h6c517f8_0
hdf5=1.10.6=h10fe05b_1
icu=58.2=h0a44026_3
idna=3.4=py310hecd8cb5_0
importlib-metadata=4.11.3=py310hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.15.2=py310hecd8cb5_0
ipython=8.6.0=py310hecd8cb5_0
jedi=0.18.1=py310hecd8cb5_1
jinja2=3.1.2=pypi_0
jpeg=9e=hca72f7f_0
jsonschema=4.17.1=pypi_0
jupyter_client=7.4.7=py310hecd8cb5_0
jupyter_core=4.11.2=py310hecd8cb5_0
keras=2.9.0=py310hecd8cb5_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=py310he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcurl=7.85.0=h6dfd666_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libnghttp2=1.46.0=ha29bfda_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.10.0=h0a4fc7d_0
libtiff=4.4.0=h2cd0358_2
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.3=h23ab428_1
markdown=3.3.4=py310hecd8cb5_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=py310hecd8cb5_0
matplotlib-base=3.5.3=py310hfb0c5b7_0
matplotlib-inline=0.1.6=py310hecd8cb5_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
multidict=6.0.2=py310hca72f7f_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.5=py310hecd8cb5_0
numexpr=2.8.4=py310h9638375_0
numpy=1.22.3=py310hdcd3fac_0
numpy-base=1.22.3=py310hfd2de13_0
oauthlib=3.2.1=py310hecd8cb5_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1s=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.1=py310h3ea8b11_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hde71d04_1
pip=22.2.2=py310hecd8cb5_0
plotly=5.11.0=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py310he9d5cce_0
psutil=5.9.0=py310hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=10.0.1=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.0=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py310hecd8cb5_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310hecd8cb5_0
pyrsistent=0.19.2=pypi_0
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.8=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd3eb1b0_0
pytz=2022.1=py310hecd8cb5_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310he9d5cce_0
re2=2022.04.01=he9d5cce_0
readline=8.2=hca72f7f_0
requests=2.28.1=py310hecd8cb5_0
requests-oauthlib=1.3.0=py_0
rich=12.6.0=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.7.3=py310h0a4c7d7_2
semver=2.13.0=pypi_0
setuptools=65.5.0=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
snappy=1.1.9=he9d5cce_0
sqlite=3.40.0=h880c91c_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.15.1=pypi_0
tenacity=8.1.0=pypi_0
tensorboard=2.9.0=py310hecd8cb5_0
tensorboard-data-server=0.6.0=py310h55d743e_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.9.1=eigen_py310ha64145f_1
tensorflow-base=2.9.1=eigen_py310h36d2db2_1
tensorflow-estimator=2.9.0=py310hecd8cb5_0
termcolor=2.1.0=py310hecd8cb5_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=py310hca72f7f_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.3.0=py310hecd8cb5_0
tzdata=2022.6=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.12=py310hecd8cb5_0
validators=0.20.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.14.1=py310hca72f7f_0
xz=5.2.6=hca72f7f_0
yaml=0.2.5=haf1e3a3_0
yarl=1.8.1=py310hca72f7f_0
zeromq=4.3.4=h23ab428_0
zipp=3.8.0=py310hecd8cb5_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
